Move Trace, TraceSummary from common to query package
