+++ zipkin-common/src/main/scala/com/twitter/zipkin/query/TraceCombo.scala
@@ -15,8 +15,6 @@
  */
 package com.twitter.zipkin.query
 
-import com.twitter.zipkin.common.{Trace, TraceSummary}
-
 object TraceCombo {
   def apply(trace: Trace): TraceCombo = {
     TraceCombo(trace, TraceSummary(trace), TraceTimeline(trace), trace.toSpanDepths)
